feat: introduce model file
