+++ projects/storefrontapp-e2e-cypress/cypress/helpers/accessibility/tabbing-order/address-book.ts
@@ -1,4 +1,5 @@
-import { checkAllElements, TabElement } from '../tabbing-order';
+import { checkAllElements } from '../tabbing-order';
+import { TabElement } from '../tabbing-order.model';
 
 export function setupForAddressBookTests() {
   addAddress();
