+++ projects/storefrontapp-e2e-cypress/cypress/helpers/accessibility/tabbing-order/cart.ts
@@ -1,4 +1,5 @@
-import { checkAllElements, TabElement, testProductUrl } from '../tabbing-order';
+import { checkAllElements, testProductUrl } from '../tabbing-order';
+import { TabElement } from '../tabbing-order.model';
 
 export function cartTabbingOrder(config: TabElement[]) {
   addCartItemsAndLoadCart();
