+++ projects/storefrontapp-e2e-cypress/cypress/helpers/accessibility/tabbing-order/reset-password.ts
@@ -1,8 +1,5 @@
-import {
-  getFormFieldByValue,
-  checkAllElements,
-  TabElement,
-} from '../tabbing-order';
+import { getFormFieldByValue, checkAllElements } from '../tabbing-order';
+import { TabElement } from '../tabbing-order.model';
 
 export function forgotPasswordTabbingOrder(config: TabElement[]) {
   cy.visit('/login/forgot-password');
