+++ projects/storefrontapp-e2e-cypress/cypress/helpers/accessibility/tabbing-order/update-email.ts
@@ -1,8 +1,5 @@
-import {
-  getFormFieldByValue,
-  checkAllElements,
-  TabElement,
-} from '../tabbing-order';
+import { getFormFieldByValue, checkAllElements } from '../tabbing-order';
+import { TabElement } from '../tabbing-order.model';
 
 export function updateEmailTabbingOrder(config: TabElement[]) {
   cy.visit('/my-account/update-email');
