+++ projects/storefrontapp-e2e-cypress/cypress/helpers/accessibility/tabbing-order/wishlist.ts
@@ -1,4 +1,5 @@
-import { checkAllElements, TabElement } from '../tabbing-order';
+import { checkAllElements } from '../tabbing-order';
+import { TabElement } from '../tabbing-order.model';
 
 export function wishlistTabbingOrder(config: TabElement[]) {
   cy.server();
